Update metodos de libros autores idioma y estadistica implementado

+++ src/main/java/com/literalura/challenge/model/DatosLibro.java
@@ -10,7 +10,7 @@ public record DatosLibro(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutor> autor,
         @JsonAlias("languages") List<String> idiomas,
-        @JsonAlias("download_count") Integer numeroDeDescargas
+        @JsonAlias("download_count") Double numeroDeDescargas
 
 ) {
 }
